feat: load stylable.config.cjs as well (#2573)
to better support projects that move to "type": "module" (in package.json), allow stylable configuration file to be named stylable.config.cjs (exporting a commonjs module).

unfortunatly, stylable cannot currently support mjs config. the config loading function is sync, and we would need to import() that file if we want to support mjs.